refactor(direct-control): extract session health wrapper
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves checkCiv7DirectControlHealth into src/session/health.ts while keeping the public facade export in index.ts and leaving waitForCiv7DirectControl plus Tuner-ready waiting in the facade.

This preserves endpoint discovery composition, no-state/state-missing/unavailable classification, selected-state reporting, typed error wrapping, and existing session test ownership. It does not define telemetry/correlation contracts, public procedure schemas, semantic CLI projection, AI ingestion, hotseat runtime proof, or Effect/oRPC procedure-core work. Task 2.9.4 matrix-row acceptance remains pending.

Validation:
- bun run --cwd packages/civ7-direct-control test -- session.test.ts public-api.test.ts
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- git diff --cached --check

Local package/source relocation proof only; no runtime, live-game, AI-on-hotseat, telemetry, semantic CLI, or procedure-core proof claimed.
